pyproject.toml
[project]
name = "olympus-mcp"
version = "0.1.0"
description = "Olympus MCP Server — Aether Agents orchestrator. Manages Daimon lifecycle and inter-agent communication via ACP."
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "agent-client-protocol>=0.1.0",
    "pyyaml>=6.0",
]

[project.scripts]
olympus = "olympus.server:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
olympus = ["py.typed"]